requirements.txt
flask==3.0.2
pillow==10.2.0
numpy==1.26.4
ultralytics==8.1.28
torch==2.2.0
torchvision==0.17.0
opencv-python==4.9.0.80 